requirements.txt
pylint
pyyaml
setuptools
